+++ causeway-client/src/main/java/org/jboss/pnc/causewayclient/remotespi/NpmBuild.java
@@ -52,7 +52,7 @@ public class NpmBuild extends Build {
             String buildName,
             String buildVersion,
             String externalBuildSystem,
-            int externalBuildID,
+            Long externalBuildID,
             String externalBuildURL,
             Date startTime,
             Date endTime,
